Conflicts:
	src/db/create_art_query.cc
	src/db/create_art_query.h
	src/db/create_art_result.cc
	src/db/create_ng_query.cc
	src/db/delete_art_query.cc
	src/db/delete_art_query.h
	src/db/delete_ng_query.cc
	src/db/delete_ng_query.h
	src/db/get_art_query.cc
	src/db/list_art_query.cc
	src/db/list_art_query.h
	src/db/list_art_result.cc
	src/db/list_art_result.h
	src/db/list_ng_result.h
	src/db/query.h
	src/db/result.h
	src/server/server_main.cc
	src/test/test_db.cc

+++ src/db/get_art_query.h
@@ -8,12 +8,12 @@
 namespace db {
 	class GetArtQuery : public Query {
 	public:
-		GetArtQuery(int ng_id, int art_id) : newsgroup_id(ng_id), article_id(art_id) {}
+		GetArtQuery(size_t ng_id, size_t art_id) : newsgroup_id(ng_id), article_id(art_id) {}
 		Result *getResult(Database &db);
 		void send(net::MessageHandler &mh);
 	private:
-		int newsgroup_id;
-		int article_id;
+		size_t newsgroup_id;
+		size_t article_id;
 	};
 
 }
