+++ src/db/list_ng_result.cc
@@ -22,7 +22,7 @@ namespace db {
 	}
 
 	void ListNGResult::toString(ostream &out) const {
-		pair<int, string> current_pair;
+		pair<size_t, string> current_pair;
 		for (size_t i = 0; i < news_groups.size(); ++i) {
 			current_pair = news_groups[i];
 			out << current_pair.first << ". " << current_pair.second << endl;
